Adds dockerized express app
Also adding instructions to get setup with docker.

Why use express?

Because I think it would be cool to shallow clone repos that are most
recent from a GitHub user's `events` stream and count the lines of code
per langauge written.  That would be way more accurate data IMO.
Nevertheless, the cloning will have to write to the file system, thus a
server is required.